chore: update structure of `src/` directory for improved modularity rename `App.jsx` and `App.css` to `Home.jsx` and `Home.css`, respectively move `Home.jsx` and `Tour.jsx` to `components/pages/` directory move `Home.css` to `styles/pages/` directory add `nav/` and `sections/` directories to `components/` and `styles/` move `NavBar.jsx` and `NavLink.jsx` to `components/nav/` move `NavBar.css` and `NavLink.css` to `styles/nav` move  `About.jsx`, `Amenities.jsx`, `Booking.jsx`, `Faq.jsx`, `Hero.jsx`, `LocalAreaMap.jsx`, `Location.jsx` and `Testimonials.jsx` to `components/sections/` move  `About.css`, `Amenities.css`, `Booking.css`, `Faq.css`, `Hero.css`, `LocalAreaMap.css`, `Location.css` and `Testimonials.css` to `styles/sections/` add `sections/amenities/` directory to `components/` and `styles/` move `AmenitiesList.jsx`, `Amenity.jsx` and `FeaturedAmenity.jsx` to `components/sections/amenities/` move `AmenitiesList.css` and `Amenity.css` to `components/sections/amenities/`
