+++ src/assets/icons/bed.png
