+++ src/assets/images/cottage/owl-bedroom.webp
