+++ src/components/sections/amenities/FeaturedAmenity.jsx
@@ -1,5 +1,5 @@
 import PropTypes from "prop-types";
-import Card from "./Card";
+import Card from "../../shared/Card";
 
 export default function FeaturedAmenity({ iconSrc, heading, content }) {
 	return (
