--- src/data/copy.json
@@ -1,34 +1,0 @@
-{
-	"about": [
-		{
-			"id": "0",
-			"content": "A traditional stone cottage built in 1868, Owl Cottage features a large, sociable kitchen/dining/living area with wood burner, master bedroom with king size bed, second bedroom with bunk beds, shower room, enclosed back garden, off-street parking and stunning views over the forest and mountains opposite."
-		},
-		{
-			"id": "1",
-			"content": "The well-equipped kitchen features a breakfast bar, dishwasher, electric range cooker, fridge, freezer, Nespresso machine and washer/dryer. Plenty of storage space and a wide range of kitchen equipment makes this great area if you choose to cook on your holiday."
-		}
-	],
-	"location": [
-		{
-			"id": "0",
-			"content": "A fantastic central point to explore the Highlands and beyond."
-		},
-		{
-			"id": "1",
-			"content": "Loch Oich is just five minutes away. Not as well known as Loch Ness but a real hidden treasure. The Great Glen Way runs along the loch shore and its flat, level path here makes it highly accessible for walking and cycling. Very dog friendly with no nearby roads or livestock. Park at Bridge of Oich to access the loch."
-		},
-		{
-			"id": "2",
-			"content": "The Caledonian Canal from Bridge of Oich to Fort Augustus is another flat, accessible walk, suitable for walkers, cyclists and dogs. Follow the quiet footpath through beautiful forestry to reach the famous Fort Augustus Locks and the start of Loch Ness. You can hire cycles here, try a boat trip on the loch, or for the more adventurous, several companies offer kayaking, canoeing and white water adventures."
-		},
-		{
-			"id": "3",
-			"content": "Many more wonderful walks are within easy reach - at all levels from short, flat paths to challenging Munros. And the ultimate challenge of Ben Nevis is just 40 minutes away."
-		},
-		{
-			"id": "4",
-			"content": "The Isle of Skye is an hour's drive west, and Inverness is an hour north."
-		}
-	]
-}
